# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Singers catalog API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/albums": {
            "get": {
                "tags": [
                    "Albums"
                ],
                "summary": "Album list",
                "operationId": "9d34167d8709928c2b4c7b9a507db3ac",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Album name"
                                            },
                                            "release_date": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "2012-12-12"
                                            },
                                            "singer_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "2024-11-18T16:58:51.000000Z"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "2024-11-18T17:09:39.000000Z"
                                            },
                                            "songs": {
                                                "type": "array",
                                                "items": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Song name"
                                                        },
                                                        "created_at": {
                                                            "type": "string",
                                                            "format": "date",
                                                            "example": "2024-11-18T16:58:51.000000Z"
                                                        },
                                                        "updated_at": {
                                                            "type": "string",
                                                            "format": "date",
                                                            "example": "2024-11-18T17:09:39.000000Z"
                                                        },
                                                        "pivot": {
                                                            "properties": {
                                                                "album_id": {
                                                                    "type": "integer",
                                                                    "example": 1
                                                                },
                                                                "song_id": {
                                                                    "type": "integer",
                                                                    "example": 1
                                                                },
                                                                "track_number": {
                                                                    "type": "integer",
                                                                    "example": 1
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "singer": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Singer name"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2024-11-18T16:58:51.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2024-11-18T17:09:39.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Albums"
                ],
                "summary": "Create album",
                "operationId": "e14d18c13dc55c0e7bd086e9842297cc",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "Album name"
                                            },
                                            "release_date": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "2012-12-12"
                                            },
                                            "singer_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "song_ids": {
                                                "type": "array",
                                                "items": {
                                                    "properties": {
                                                        "singer_id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "track_number": {
                                                            "type": "integer",
                                                            "example": 1
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Album name"
                                        },
                                        "release_date": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2012-12-12"
                                        },
                                        "singer_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2024-11-18T16:58:51.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2024-11-18T17:09:39.000000Z"
                                        },
                                        "songs": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Song name"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2024-11-18T16:58:51.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2024-11-18T17:09:39.000000Z"
                                                    },
                                                    "pivot": {
                                                        "properties": {
                                                            "album_id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "song_id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "track_number": {
                                                                "type": "integer",
                                                                "example": 1
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "singer": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Singer name"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2024-11-18T16:58:51.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2024-11-18T17:09:39.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}